audio-editor/backend/requirements.txt
# API and server
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiofiles>=23.1.0
pydub>=0.25.1
openai-whisper>=20231117
numpy>=1.24.3
torch>=2.0.0
tqdm>=4.65.0
flask==2.1.2
flask-cors==3.0.10
werkzeug==2.1.2
python-dotenv==0.19.0
scipy==1.10.1 